move gradients into helpers
