+++ client/src/Components/TeamDashboard/Chart.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { Pie } from 'react-chartjs-2'
-import { colors, warnColors } from '../gradients'
+import { colors, warnColors } from '../helpers'
 
 export default class Chart extends Component {
   render () {
