+++ client/src/Components/helpers.js
@@ -1,3 +1,17 @@
+import _ from 'lodash'
+
+export function calcSalaryCap(players) {
+  const numTeams = 8
+  const salaries = _.map(players, (p) => p.salary)
+  return _.sum(salaries) / numTeams * 1.01
+}
+
+export function calcSalaryFloor(players) {
+  const numTeams = 8
+  const salaries = _.map(players, (p) => p.salary)
+  return _.sum(salaries) / numTeams * 0.99
+}
+
 export const colors = [
   '#E5F5E0',
   '#D7EDD4',
